# node-ouput/swagger.yaml
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "The Social Media",
    "description" : "API Documentation for The Social Media App",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:3000/api/v1"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/room/create" : {
      "post" : {
        "tags" : [ "Rooms" ],
        "summary" : "Create a new room",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/room_create_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Room created successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/room/{roomId}/createMessage" : {
      "post" : {
        "tags" : [ "Rooms" ],
        "summary" : "Create a new message in a room",
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "description" : "The room ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/roomId_createMessage_body"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Message created successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/room/{roomId}/messages" : {
      "get" : {
        "tags" : [ "Rooms" ],
        "summary" : "Get all messages in a room",
        "parameters" : [ {
          "name" : "roomId",
          "in" : "path",
          "description" : "The room ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of messages",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/room/last-message/{userId1}/{userId2}" : {
      "get" : {
        "tags" : [ "Rooms" ],
        "summary" : "Get the last message between two users",
        "parameters" : [ {
          "name" : "userId1",
          "in" : "path",
          "description" : "The first user ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "userId2",
          "in" : "path",
          "description" : "The second user ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Last message fetched successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/room/unread-messages/{userId}" : {
      "get" : {
        "tags" : [ "Rooms" ],
        "summary" : "Get unread messages count for a user",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The user ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Unread messages count fetched successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/room/{senderId}/{receiverId}" : {
      "get" : {
        "tags" : [ "Rooms" ],
        "summary" : "Get messages between two users",
        "parameters" : [ {
          "name" : "senderId",
          "in" : "path",
          "description" : "The sender ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "receiverId",
          "in" : "path",
          "description" : "The receiver ID",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A list of messages",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/savedposts" : {
      "get" : {
        "tags" : [ "Saved Posts" ],
        "summary" : "Get all saved posts for a user",
        "responses" : {
          "200" : {
            "description" : "A list of saved posts",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SavedPost"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/savedposts/save/{postId}" : {
      "post" : {
        "tags" : [ "Saved Posts" ],
        "summary" : "Save a post",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "description" : "The post ID to save",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Post saved successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/savedposts/unsave/{postId}" : {
      "delete" : {
        "tags" : [ "Saved Posts" ],
        "summary" : "Unsave a post",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "description" : "The post ID to unsave",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Post unsaved successfully"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Message" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109cd"
          },
          "sender" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109cb"
          },
          "receiver" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109cc"
          },
          "content" : {
            "type" : "string",
            "example" : "Hello!"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "seen" : {
            "type" : "boolean",
            "example" : false
          },
          "room" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109ca"
          }
        }
      },
...
      "SavedPost" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109ca"
          },
          "user" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109cb"
          },
          "post" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109cc"
          },
          "savedAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2021-06-24T14:48:00Z"
          }
        }
      },
...
      "room_create_body" : {
        "type" : "object",
        "properties" : {
          "user1Id" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109ca"
          },
          "user2Id" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109cb"
          }
        }
      },
      "roomId_createMessage_body" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string",
            "example" : "Hello!"
          },
          "senderId" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109ca"
          },
          "receiverId" : {
            "type" : "string",
            "example" : "60d0fe4f5311236168a109cb"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}